Restrict some competitions to members only #10919
Some competitions should be only available to NUCATS Members.
This change adds a flag to the Project model and if this
membership_required flag is set to true, the application
checks myNUCATS to see if the current logged in user is
a member. If not the user is redirected to a
`membership_required` page letting them know of this
requirement.