Reset of public repository.
There were some sensitive details inadvertently added to the repository
some time ago, and have somehow managed to stay in the repo for a long
time.

Over the last couple of months we have quietly removed those details,
and have decided to now reset the repository to the current version, so
that the history of those changes is gone.