created rake tasks to create reports on projects. ASSIST-142
There was already a csv generating task, but it created many csvs. I split this out into separate functions so that the csvs could be generated separately. Once separated, I created the new tasks :herzog, :new_projects, and :updated_projects. The task :herzog asked for more fields than were in the original :projects csv. Instead of changing the original, I left that alone, but added a new task in addition to the other two which are meant to be run on a regular schedule.